apps/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI backend"
requires-python = ">=3.11"

dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.29",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.1",
  "pydantic-settings>=2.2",
  "pydantic>=2.0",
  "pydantic[email]>=2.0",
  "numpy>=1.24.0",
  "opencv-python>=4.8.0",
  "dlib>=19.24.0",
  "pgvector>=0.3.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "python-multipart>=0.0.6",
  "alembic>=1.13.0",
  "bcrypt<4.0",
  "pillow-heif>=1.1.1",
]

[tool.uv]
dev-dependencies = [
  "pytest",
  "pytest-asyncio",
  "httpx",
]
